chore: UserControllerTest에 있었던 로그인 관련 테스트를 LoginControllerTest로 이동
