+++ src/test/java/kr/codesqaud/cafe/app/user/controller/UserControllerTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.redirectedUrl;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -13,14 +12,13 @@ import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import kr.codesqaud.cafe.app.user.entity.User;
-import kr.codesqaud.cafe.app.user.repository.UserRepository;
-import kr.codesqaud.cafe.app.user.controller.dto.UserLoginRequest;
 import kr.codesqaud.cafe.app.user.controller.dto.UserResponse;
 import kr.codesqaud.cafe.app.user.controller.dto.UserSavedRequest;
+import kr.codesqaud.cafe.app.user.entity.User;
+import kr.codesqaud.cafe.app.user.repository.UserRepository;
+import kr.codesqaud.cafe.errors.errorcode.UserErrorCode;
 import kr.codesqaud.cafe.errors.response.ErrorResponse;
 import kr.codesqaud.cafe.errors.response.ErrorResponse.ValidationError;
-import kr.codesqaud.cafe.errors.errorcode.UserErrorCode;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -186,115 +184,16 @@ class UserControllerTest {
     }
 
     @Test
-    @DisplayName("올바른 유저아이디와 올바른 패스워드가 주어지고 로그일할때 세션 저장소에 회원정보가 저장되는지 테스트")
-    public void login_success() throws Exception {
-        //given
-        String userId = "yonghwan1107";
-        String password = "yonghwan1107";
-        String url = "/login";
-        UserLoginRequest dto = new UserLoginRequest(userId, password);
-
-        //when
-        mockMvc.perform(post(url)
-                .session(session)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(toJSON(dto)))
-            .andExpect(redirectedUrl("/"));
-        //then
-        UserResponse user = (UserResponse) session.getAttribute("user");
-        assertThat(user.getUserId()).isEqualTo(userId);
-        assertThat(user.getName()).isEqualTo("김용환");
-        assertThat(user.getEmail()).isEqualTo("yonghwan1107@naver.com");
-    }
-
-    @Test
-    @DisplayName("올바른 유저아이디와 틀린 비밀번호가 주어지고 로그인할때 비밀번호 불일치로 에러 응답을 받는지 테스트")
-    public void login_fail1() throws Exception {
-        //given
-        String userId = "yonghwan1107";
-        String password = "useaweiofjaw";
-        String url = "/login";
-        UserLoginRequest dto = new UserLoginRequest(userId, password);
-        //when
-        String jsonErrorResponse = mockMvc.perform(post(url)
-                .session(session)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(toJSON(dto)))
-            .andExpect(status().isBadRequest())
-            .andReturn().getResponse().getContentAsString(StandardCharsets.UTF_8);
-        //then
-        ErrorResponse actual = objectMapper.readValue(jsonErrorResponse, ErrorResponse.class);
-        ErrorResponse expected = new ErrorResponse(
-            UserErrorCode.NOT_MATCH_LOGIN.name(),
-            HttpStatus.BAD_REQUEST,
-            UserErrorCode.NOT_MATCH_LOGIN.getMessage(),
-            null);
-        assertThat(actual).isEqualTo(expected);
-    }
-
-    @Test
-    @DisplayName("부적절한 입력 형식의 아이디와 패스워드가 주어지고 로그인할때 에러 응답을 받는지 테스트")
-    public void login_fail2() throws Exception {
-        //given
-        String userId = "";
-        String password = "";
-        String url = "/login";
-        UserLoginRequest dto = new UserLoginRequest(userId, password);
-        //when
-        String jsonErrors = mockMvc.perform(post(url)
-                .session(session)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(toJSON(dto)))
-            .andExpect(status().isBadRequest())
-            .andReturn().getResponse().getContentAsString(StandardCharsets.UTF_8);
-        //then
-        List<ValidationError> errors = new ArrayList<>();
-        errors.add(new ValidationError("password", "8~16자 영문 대 소문자, 숫자, 특수문자만 사용 가능합니다."));
-        errors.add(new ValidationError("userId", "5~20자의 영문 소문자, 숫자와 특수기호(_),(-)만 사용 가능합니다."));
-
-        ErrorResponse errorResponse = objectMapper.readValue(jsonErrors, ErrorResponse.class);
-        Assertions.assertThat(errorResponse.getHttpStatus()).isEqualTo(HttpStatus.BAD_REQUEST);
-        Assertions.assertThat(errorResponse.getErrorMessage()).isEqualTo("유효하지 않은 입력 형식입니다.");
-        Assertions.assertThat(errorResponse.getErrors()).containsAll(errors);
-    }
-
-    @Test
-    @DisplayName("서버에 없는 회원 아이디가 주어지고 로그인할때 에러 응답을 받는지 테스트")
-    public void login_fail3() throws Exception {
-        //given
-        String userId = "user10";
-        String password = "user10user10";
-        String url = "/login";
-        UserLoginRequest dto = new UserLoginRequest(userId, password);
-        //when
-        String jsonErrorResponse = mockMvc.perform(post(url)
-                .session(session)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(toJSON(dto)))
-            .andExpect(status().isBadRequest())
-            .andReturn().getResponse().getContentAsString(StandardCharsets.UTF_8);
-        //then
-        ErrorResponse actual = objectMapper.readValue(jsonErrorResponse, ErrorResponse.class);
-        ErrorResponse expected = new ErrorResponse(
-            UserErrorCode.NOT_MATCH_LOGIN.name(),
-            HttpStatus.BAD_REQUEST,
-            UserErrorCode.NOT_MATCH_LOGIN.getMessage(),
-            null);
-        assertThat(actual).isEqualTo(expected);
-    }
-
-    @Test
-    @DisplayName("변경된 비밀번호, 이름, 이메일이 주어지고 유저아이디가 주어질때 회원정보 수정이 되는지 테스트")
+    @DisplayName("비밀번호, 이름, 이메일이 주어지고 유저아이디가 주어질때 회원정보 수정이 되는지 테스트")
     public void update_success() throws Exception {
         //given
         String userId = "yonghwan1107";
-        String modifiedPassword = "yonghwan1234";
+        String password = "yonghwan1107";
         String modifiedName = "홍길동";
         String modifiedEmail = "yonghwan1234@naver.com";
         Long id = userRepository.findByUserId(userId).orElseThrow().getId();
         String url = "/users/" + id + "/update";
-        UserSavedRequest dto = new UserSavedRequest(userId, modifiedPassword, modifiedName,
-            modifiedEmail);
+        UserSavedRequest dto = new UserSavedRequest(userId, password, modifiedName, modifiedEmail);
         //when
         mockMvc.perform(put(url)
                 .contentType(MediaType.APPLICATION_JSON)
@@ -304,7 +203,7 @@ class UserControllerTest {
         //then
         User actual = userRepository.findByUserId(userId).orElseThrow();
         assertThat(actual.getName()).isEqualTo(modifiedName);
-        assertThat(actual.getPassword()).isEqualTo(modifiedPassword);
+        assertThat(actual.getPassword()).isEqualTo(password);
         assertThat(actual.getEmail()).isEqualTo(modifiedEmail);
     }
 
@@ -333,46 +232,6 @@ class UserControllerTest {
             UserErrorCode.ALREADY_EXIST_EMAIL.name(),
             HttpStatus.CONFLICT,
             UserErrorCode.ALREADY_EXIST_EMAIL.getMessage(),
-            null);
-        assertThat(actual).isEqualTo(expected);
-    }
-
-    @Test
-    @DisplayName("비밀번호가 맞는지 테스트")
-    public void password_success() throws Exception {
-        //given
-        Long id = userRepository.findByUserId("yonghwan1107").orElseThrow().getId();
-        String url = "/password/" + id;
-        String password = "yonghwan1107";
-        UserSavedRequest dto = new UserSavedRequest(null, password, null, null);
-        //when & then
-        mockMvc.perform(post(url)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(toJSON(dto)))
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl("/user/form/" + id));
-    }
-
-    @Test
-    @DisplayName("비밀번호가 틀려서 에러 메시지를 응답 받는지 테스트")
-    public void password_fail() throws Exception {
-        //given
-        Long id = userRepository.findByUserId("yonghwan1107").orElseThrow().getId();
-        String url = "/password/" + id;
-        String password = "awioefjoawiefj";
-        UserSavedRequest dto = new UserSavedRequest(null, password, null, null);
-        //when
-        String jsonErrorResponse = mockMvc.perform(post(url)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(toJSON(dto)))
-            .andExpect(status().isBadRequest())
-            .andReturn().getResponse().getContentAsString(StandardCharsets.UTF_8);
-        //then
-        ErrorResponse actual = objectMapper.readValue(jsonErrorResponse, ErrorResponse.class);
-        ErrorResponse expected = new ErrorResponse(
-            UserErrorCode.NOT_MATCH_PASSWORD.name(),
-            HttpStatus.BAD_REQUEST,
-            UserErrorCode.NOT_MATCH_PASSWORD.getMessage(),
             null);
         assertThat(actual).isEqualTo(expected);
     }
